[FIX] mrp_account: wrong company on analytic line when validating MO
When marking as done a MO, we use SUPERUSER to create analytic line
as an MRP user does not always have access to analytic items. There
is however a default value for company analytic line defined to the
user's context company, so we need to use the current company of the
real user instead.

closes odoo/odoo#39209

X-original-commit: 14aade9341bf9d61cd21de948755fc339ae512ea
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>