[FIX] stock: fix delivery slip serial number creation
In deliveries slips, you can directly create a new serial number.
If you have multiples companies, a field "company" appears.
But if you have only one company, it doesn't appears and there is
no default value for it.

As the field is required, it returns an error "Company field required".

Before this commit:

    - The company_id field in stock_production_lot model does not
      have a default value.

After this commit:

    - The company_id field has a default value which is the user
      current company id.

OPW-2083099

closes odoo/odoo#39170

X-original-commit: 7b6d87519d3ba50688f9f49442cc3814ee427816
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>