[FIX] website: GA key
`_` is not available when loading this part of the template, leading to
an error on `_.str.trim()`.

We can simply use `trim()` to do so. By the way, `google_analytics_key`
is now trimmed automatically when adding it in the settings, so this is
only useful for migrated databases.

Fixes #38891
opw-2089296

closes odoo/odoo#39269

X-original-commit: a492047e9a064beecbe0611c3f7aef5172406eb1
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>