[FIX] point_of_sale: no read access to bank statement
When installing point_of_sale, the admin user only belongs to
group_account_invoice and not on group_account_user. Because of this,
the record rule define in point_of_sale that allows access to
account.bank.statement and account.bank.statement.line models does not
work as intended. This commit now allows read access to the said account
models.

Task-id: 2083597
X-original-commit: b521605c166ba03079bc15899a7dc3129326c98c