[ADD] l10n_ar: activate purhcase suffered perceptions
This ones were inactive by default, we think that need to be active in
order to be able to be added to vendor bills.

Also change the amount from 0.0 to 1.0 of all the perception tax
templates to make it appears in the manual taxes widget when they are
addes to the invoice/bills lines.

closes odoo/odoo#39400

X-original-commit: ff70835db4f6354de62a00dc717682c2835a105c
Signed-off-by: Josse Colpaert <jco@openerp.com>