[REF] stock: immediate and backorder wizard multi
followup of rev [0]

If these wizards are called on multiple pickings, display the list of
the pickings that could be impacted and allow to select which one should
be impacted.

We also adapt the sanity checks at the start of `button_validte` in
order to specify the concerned pickings if needed. We do not enable the
multi behavior for batch at the moment, so it's only enabled for the
validate multi in the list view.

[0] 08949d6388a3e9d203c51061bd7ed7616ccb2dbd

task-2069646

closes odoo/odoo#41497

X-original-commit: 3c182a0c07b5cd2d6921798e8351c082274b465c
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>