[FIX] stock: picking: immediate and state
Issue: create an immediate transfer, add a move and a done quantity, the
picking is in waiting state instead of ready.

Due to rev[0], the moves added in an immediate transfer have an initial
demand. After auto-confirming them, their state were 'confirmed' and not
'assigned'.

Solution: we force the state to assigned

[0] 7a0c30f7277e24a9dd1e2821bfc8a26147bd198c

closes odoo/odoo#39492

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>